+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Anticipated release: December 21, 2020
 #### 🚀 New features
 
 - Update endpoint for affiliations to filter by status ([#2682])
+- Adds State Admin panel ([#2583])
 
 #### 🐛 Bugs fixed
 
@@ -17,3 +18,4 @@ Anticipated release: December 21, 2020
 See our [release history](https://github.com/CMSgov/eAPD/releases)
 
 [#2682]: https://github.com/CMSgov/eAPD/issues/2682
+[#2583]: https://github.com/CMSgov/eAPD/issues/2583